backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
openai
langchain
langchain-community
langchain-openai
tiktoken
PyPDF2
faiss-cpu
numpy
